apps/api/requirements.txt
django>=5.1,<5.2
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-filter>=24.0,<25.0
psycopg[binary]>=3.2,<4.0
redis>=5.0,<6.0
celery>=5.4,<6.0
structlog>=24.1,<26.0
django-structlog>=8.0,<11.0
pytest>=8.0,<9.0
pytest-django>=4.9,<5.0
factory-boy>=3.3,<4.0
requests>=2.32,<3.0
drf-nested-routers>=0.93.4,<1.0
ruff>=0.8,<1.0
